+++ Sources/BraintreeVenmo/BTVenmoRequest.m
@@ -1,0 +1,9 @@
+#ifdef COCOAPODS
+#import <Braintree/BTVenmoRequest.h>
+#else
+#import <BraintreeVenmo/BTVenmoRequest.h>
+#endif
+
+@implementation BTVenmoRequest
+
+@end